requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via
    #   aiohttp
    #   langchain
attrs==22.2.0
    # via aiohttp
authlib==1.2.0
    # via weaviate-client
backoff==2.2.1
    # via posthog
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   nltk
    #   uvicorn
cmake==3.26.4
    # via triton
cryptography==41.0.1
    # via authlib
dataclasses-json==0.5.7
    # via langchain
decorator==5.1.1
    # via validators
fastapi==0.95.0
    # via
    #   -r requirements.in
    #   sentry-sdk
filelock==3.12.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.5.0
    # via huggingface-hub
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
grpcio==1.54.2
    # via
    #   grpcio-tools
    #   qdrant-client
grpcio-tools==1.54.2
    # via qdrant-client
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.17.2
    # via httpx
httpx[http2]==0.24.1
    # via qdrant-client
huggingface-hub==0.15.1
    # via
    #   sentence-transformers
    #   transformers
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.2
    # via torch
joblib==1.2.0
    # via
    #   nltk
    #   scikit-learn
langchain==0.0.193
    # via -r requirements.in
langchainplus-sdk==0.0.4
    # via langchain
lit==16.0.5.post0
    # via triton
markupsafe==2.1.3
    # via jinja2
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
monotonic==1.6
    # via posthog
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.1
    # via torch
nltk==3.8.1
    # via sentence-transformers
numexpr==2.8.4
    # via langchain
numpy==1.23.5
    # via
    #   langchain
    #   numexpr
    #   qdrant-client
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
openai==0.27.2
    # via -r requirements.in
openapi-schema-pydantic==1.2.4
    # via langchain
packaging==23.0
    # via
    #   huggingface-hub
    #   marshmallow
    #   transformers
pdpyras==4.5.2
    # via -r requirements.in
pillow==9.5.0
    # via torchvision
portalocker==2.7.0
    # via qdrant-client
posthog==2.5.0
    # via -r requirements.in
protobuf==4.23.2
    # via grpcio-tools
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via
    #   fastapi
    #   langchain
    #   langchainplus-sdk
    #   openapi-schema-pydantic
    #   qdrant-client
python-dateutil==2.8.2
    # via posthog
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0
    # via
    #   huggingface-hub
    #   langchain
    #   transformers
qdrant-client==1.2.0
    # via -r requirements.in
regex==2023.3.23
    # via
    #   nltk
    #   tiktoken
    #   transformers
requests==2.28.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   langchain
    #   langchainplus-sdk
    #   openai
    #   pdpyras
    #   posthog
    #   tiktoken
    #   torchvision
    #   transformers
    #   weaviate-client
safetensors==0.3.1
    # via transformers
scikit-learn==1.2.2
    # via sentence-transformers
scipy==1.10.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.2.2
    # via -r requirements.in
sentencepiece==0.1.99
    # via sentence-transformers
sentry-sdk[fastapi]==1.19.1
    # via -r requirements.in
six==1.16.0
    # via
    #   posthog
    #   python-dateutil
slack-bolt==1.17.0
    # via -r requirements.in
slack-sdk==3.20.2
    # via slack-bolt
smmap==5.0.0
    # via gitdb
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.8
    # via langchain
starlette==0.26.1
    # via fastapi
sympy==1.12
    # via torch
tenacity==8.2.2
    # via
    #   langchain
    #   langchainplus-sdk
threadpoolctl==3.1.0
    # via scikit-learn
tiktoken==0.3.3
    # via -r requirements.in
tokenizers==0.13.3
    # via transformers
torch==2.0.1
    # via
    #   sentence-transformers
    #   torchvision
    #   triton
torchvision==0.15.2
    # via sentence-transformers
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   nltk
    #   openai
    #   sentence-transformers
    #   transformers
    #   weaviate-client
transformers==4.30.0
    # via sentence-transformers
triton==2.0.0
    # via torch
typing-extensions==4.5.0
    # via
    #   huggingface-hub
    #   pydantic
    #   qdrant-client
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==1.26.15
    # via
    #   pdpyras
    #   qdrant-client
    #   requests
    #   sentry-sdk
uvicorn==0.21.1
    # via -r requirements.in
validators==0.20.0
    # via weaviate-client
weaviate-client==3.19.2
    # via -r requirements.in
wheel==0.40.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
